requirements.txt
tqdm
numpy
pandas
albumentations
timm
opencv-python
scikit-learn
pytorch-lightning
omegaconf
wandb
transformers
pyarrow
jupyter
matplotlib
torch_geometric
h5py
